Struct jsonata::value::Value

pub struct Value {
    pub pool: ValuePool,
    pub index: usize,
}

A thin wrapper around the index to a ValueKind within a ValuePool.

Values are intended to be created and dropped as needed without having any effect on the underlying data stored in the ValuePool. They are essentially pointers to nodes in the pool, that contain a reference to the pool and the index of a node in the pool.

As a Value is just an Rc and a usize, it has a Clone implementation which makes it very easy to pass around.

Fields
pool: ValuePoolindex: usize
Implementations

impl Value

pub fn new(pool: ValuePool, kind: ValueKind) -> Value

pub fn new_undefined(pool: ValuePool) -> Value

pub fn new_null(pool: ValuePool) -> Value

pub fn new_bool(pool: ValuePool, value: bool) -> Value

pub fn new_number(pool: ValuePool, value: Number) -> Value

pub fn new_string(pool: ValuePool, value: &str) -> Value

pub fn new_array(pool: ValuePool) -> Value

pub fn new_array_with_capacity(pool: ValuePool, capacity: usize) -> Value

pub fn new_object(pool: ValuePool) -> Value

pub fn new_object_with_capacity(pool: ValuePool, capacity: usize) -> Value

pub fn is_undefined(&self) -> bool

pub fn is_null(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_number(&self) -> bool

pub fn is_string(&self) -> bool

pub fn is_array(&self) -> bool

pub fn is_object(&self) -> bool

pub fn as_bool(&self) -> bool

pub fn as_f64(&self) -> f64

pub fn as_string(&self) -> String

pub fn get_member(&self, index: usize) -> Value

pub fn get_entry(&self, key: &str) -> Value

pub fn insert(&self, key: &str, kind: ValueKind)

pub fn insert_index(&self, key: &str, index: usize)

pub fn push(&self, kind: ValueKind)

Pushes a new ValueKind into the ValueKind::Array wrapped by this Value.

Panics

If the ValueKind wrapped by this Value is anot a ValueKind::Array.

pub fn push_index(&self, index: usize)

Pushes a pool index into the ValueKind::Array wrapped by this Value.

Use this if you have constructed a Value separately and now want it to be an item in an existing ValueKind::Array.

Note: This makes absolutely no attempt to a) verify the the index is valid, b) that it even came from the same ValuePool, or c) that this does not create a circular reference.

Panics

If the ValueKind wrapped by this Value is not a ValueKind::Array.

pub fn drop(self)

Removes the actual ValueKind wrapped by the Value from the pool and consumes the Value.

As Values are thin wrappers around an index in the pool, and there could be other Values referencing the indexed ValueKind, this is not in a real Drop implementation.

“Dropping” in this case actually means removing the ValueKind from the pool and putting its index on the free list.

pub fn wrap_in_array(self) -> Value

Wraps an existing value in an array by creating a new array in the pool and adding this Value’s index as the only item.

This is used extensively throughout evaluation to implement the rules of sequences, and one of the actions that has a major advantage using the pool implementation. If there wasn’t a pool, the entire input would need to be cloned which could be very costly.

pub fn members(&self) -> Members<'_>

Create an iterator over the members of an array.

Panics

If the ValueKind wrapped by this Value is not a ValueKind::Array.

pub fn entries(&self) -> Entries<'_>

Create an iterator over the entries of an object.

Panics

If the ValueKind wrapped by this Value is not a ValueKind::Object.